requirements.txt
# Web Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Mail

# Database - PostgreSQL Support
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Data Processing & AI/ML
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0

# Data Visualization
plotly==5.18.0

# Utilities
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.3

# Forecasting
fbprophet

# Werkzeug (Flask dependency)
Werkzeug==2.3.8
